Adds fetcher and filing types
This commit sets us up to be able to review filings for a given CIK. It adds basic types for filings and for the JSON representation of a CIK's data from the SEC Resful API. This basic setup should enable us to do the basic math on whether or not a Form 4 filing was late. This is also the basis for any other operations in the future on different types of forms.

The typing can probably be cleaned up or simplified in the future. Additionally there likely is some better way to handle the columnar data from the SEC to turn it into forms. I can't think of one right now however. Next step is probably to create some sort of utility to do this work.